Distinguish between Command and TimeStep Actions
hopefully making it clearer that some things (e.g. a WaitAction) can be
part of a solution command, but can't be run "at" a time step - instead
they must be expanded by the SolutionInterpreter.  On the other hand,
actions like PlaceTokenAction can be part of a solution command, and be
an action to be taken at a time step.