Characters for water on ramps should follow the standard rules.
Previously we were trying to be clever and using N to mean the region was
full of water, taking into account the capacity. Now that we've implemented
the ability to specify standing water in a rel file, this is confusing, and
it's much simpler for N to always mean 1024 units, and n to always mean 512
units (unless specified otherwise by water description lines).