Fix multiple Swing rendering problems.  Closes #452.
Only call redraw() from the Swing event loop when it is not going to be
called from the game loop, because we are paused.

Refuse to actually do the re-drawing if another thread is already doing
it.

Launch a new thread when we do choose to redraw from within the Swing
event loop.

Do call this code when the window is activated as well as when it is
resized, to prevent not re-drawing when the window has been invisible.

This prevents lots of flickering when you resize on Swing, but makes the
paused game update itself, without ending up with a huge queue of
componentResized() events all waiting to do the re-drawing, as we saw in